Move hook modules from src/components/ to src/hooks/
